+++ src/components/AccountDrawer/MiniPortfolio/NFTs/index.tsx
@@ -5,9 +5,9 @@ import { useState } from 'react'
 import InfiniteScroll from 'react-infinite-scroll-component'
 import styled from 'styled-components/macro'
 
-import { useAccountDrawer } from '..'
-import { DEFAULT_NFT_QUERY_AMOUNT } from './constants'
-import { NFT } from './NFT'
+import { useAccountDrawer } from '../..'
+import { DEFAULT_NFT_QUERY_AMOUNT } from '../constants'
+import { NFT } from './NFTItem'
 
 export default function NFTs({ account }: { account: string }) {
   const [walletDrawerOpen, toggleWalletDrawer] = useAccountDrawer()
